Refactor: renaming package to candig_rnaget
