+++ candig_rnaget/scripts/post_data.py
@@ -9,8 +9,8 @@ import requests
 import os
 import json
 
-from rnaget_service.api.operations import validate_uuid_string
-from rnaget_service.api.exceptions import AuthorizationError
+from candig_rnaget.api.operations import validate_uuid_string
+from candig_rnaget.api.exceptions import AuthorizationError
 
 
 def add_expression_file(args=None):
