+++ setup.cfg
@@ -4,6 +4,9 @@ universal = 1
 [flake8]
 exclude = docs
 
+[metadata]
+description-file = README.md
+
 [aliases]
 # Define setup.py command aliases here
 test = pytest
