+++ tests/dredd.yml
@@ -2,7 +2,7 @@ dry-run: null
 hookfiles: null
 language: python
 sandbox: false
-server: python3 -m rnaget_service --database=test.db --logfile=test.log --loglevel=INFO --port=8045
+server: python3 -m candig_rnaget --database=test.db --logfile=test.log --loglevel=INFO --port=8045
 server-wait: 30
 init: false
 custom: {}
@@ -29,5 +29,5 @@ hooks-worker-term-retry: 500
 hooks-worker-handler-host: 127.0.0.1
 hooks-worker-handler-port: 61321
 config: ./dredd.yml
-blueprint: ../rnaget_service/api/rnaget.yaml
+blueprint: ../candig_rnaget/api/rnaget.yaml
 endpoint: 'http://ga4ghdev01.bcgsc.ca:8045'
